requirements.txt
adafruit-pca9685>=1.0.1
wiringpi>=2.60.1
RPi.GPIO>=0.7.0
picamera>=1.13
smbus>=1.1.post2
